feat(tools): :sparkles: split sim_flight_kalman
- create new script only for plotting data.
- update sim_flight_kalman with current filter implementation
- update documentation
- update tools to use plotter
