+++ doc/tools/index.md
@@ -15,6 +15,7 @@ Pip requirements are installed via `pip install -r tools/requirements.txt`
 ```{toctree}
 :maxdepth: 1
 
+plot_flight_data
 sim_fetch
 sim_flight_kalman
 sweep_apogee
@@ -25,7 +26,8 @@ sync_defconfig
 
 | Script | Purpose |
 |---|---|
+| [`plot_flight_data.py`](plot_flight_data.md) | Plot recorded or simulated flight data — standalone CLI for telemetry logs, importable plotting module for `sim_flight_kalman`. |
 | [`sim_fetch.py`](sim_fetch.md) | Pull files out of a running `native_sim` Zephyr instance via the shell console. |
-| [`sim_flight_kalman.py`](sim_flight_kalman.md) | Replay simulated or recorded flights through a Python mirror of the AURORA Kalman filter and plot the result. |
+| [`sim_flight_kalman.py`](sim_flight_kalman.md) | Run a synthetic flight through a Python mirror of the AURORA Kalman filter and attitude tracker. |
 | [`sweep_apogee.py`](sweep_apogee.md) | Grid-search Kalman filter tuning parameters against recorded flights. |
 | [`sync_defconfig.py`](sync_defconfig.md) | Merge the generated Zephyr `defconfig` into a board-specific `.conf`. |
