+++ doc/tools/sim_flight_kalman.md
@@ -1,81 +1,76 @@
 # sim_flight_kalman.py
 
-Replay a simulated or recorded rocket flight through a Python mirror of the
-AURORA Kalman filter and inspect the result graphically.
+Run a synthetic rocket flight through a Python mirror of the AURORA
+Kalman filter and plot the result.
 
 ## What it does
 
-The script replicates the two-state (altitude, velocity) constant-
-acceleration Kalman filter from [`aurora/lib/filter/kalman.c`](https://github.com/AUXSPACEeV/aurora/tree/main/lib/filter/kalman.c)
-together with the body-frame gravity tracker from
-[`aurora/lib/sensor/attitude.c`](https://github.com/AUXSPACEeV/aurora/tree/main/lib/sensor/attitude.c),
-so that filter tuning and apogee detection can be analysed offline without
-flashing firmware.
+The script is a 1:1 Python port of the firmware filter pipeline. The two
+mirrored components are:
 
-Two data sources are supported:
+- The two-state (altitude, velocity) constant-acceleration Kalman filter
+  from [`aurora/lib/filter/kalman.c`](https://github.com/AUXSPACEeV/aurora/tree/main/lib/filter/kalman.c)
+  (`Filter`, `filter_init`, `filter_predict`, `filter_update`,
+  `filter_detect_apogee`).
+- The body-frame gravity tracker from
+  [`aurora/lib/sensor/attitude.c`](https://github.com/AUXSPACEeV/aurora/tree/main/lib/sensor/attitude.c)
+  (`Attitude`, `attitude_init`, `attitude_calibrate_sample`,
+  `attitude_calibrate_finish`, `attitude_update`,
+  `attitude_is_calibrated`).
 
-- **Synthetic flight (default):** a pressure signal is generated from the
-  ISA barometric formula and a body-frame IMU stream (accel + gyro) is
-  synthesized with a slow pitch-over after apogee.
-- **Recorded flight (`--flight DIR`):** the script loads telemetry from
-  `DIR/flights.influx` and the state machine trace from `DIR/state_audit`,
-  segments the log into individual flights on `ARMED → BOOST` transitions,
-  and produces one plot per flight with state transitions drawn as vertical
-  markers.
+Function names, struct layout, return semantics (0 = OK, 1 =
+gated/apogee, negative `errno` on error) and arithmetic order all track
+the C, so any tuning change made in Kconfig or in the C sources can be
+reproduced in Python by updating the mirrored constants at the top of
+the file.
 
-Plots cover pressure, altitude, vertical velocity, the body-frame gravity
-direction, the gravity-removed world-vertical acceleration, and the
-three-way apogee vote.
+The script generates its own synthetic data:
+
+- A pressure signal from the ISA barometric formula.
+- A body-frame IMU stream (accel + gyro) with a slow pitch-over after
+  apogee, so the gravity tracker actually has to do something.
+- MS5607-flavoured Gaussian noise on top of the pressure stream.
+
+These streams are fed through the filter mirrors at 50 Hz and the
+results are handed to [`plot_flight_data.py`](plot_flight_data.md),
+which produces the seven-panel plot shown below.
+
+```{note}
+Plotting is delegated to `plot_flight_data` — `sim_flight_kalman.py` no
+longer plots anything itself. To plot recorded telemetry from a flight
+log, use `plot_flight_data.py --flight DIR` directly.
+```
 
 ## Usage
 
 ```
-python3 tools/sim_flight_kalman.py [--flight DIR] [--show]
-                                   [--theme {light,dark,both}]
-                                   [--dt S] [--pre-boost S] [--cal S]
-                                   [--post-main S]
-                                   [--q-alt Q] [--q-vel Q] [--r-meas R]
-                                   [--debounce N]
+python3 tools/sim_flight_kalman.py [--theme {light,dark,both}]
+                                   [--show] [--title TITLE]
 ```
 
-### Data selection
-
-- `--flight DIR` — load real flight data from `DIR` (expects
-  `flights.influx` and `state_audit`). Without this flag, a synthetic
-  flight is generated.
-- `--dt` — resampling period in seconds for real flight data (default
-  `0.02`).
-- `--pre-boost` — seconds of pre-boost data to keep per flight (default
-  `10`).
-- `--cal` — length of the calibration window before BOOST (default `3`).
-- `--post-main` — seconds of data to keep after the `MAIN` transition,
-  trimming the long post-parachute tail (default `5`).
-
-### Kalman tuning
-
-The defaults match the values committed in Kconfig so runs are directly
-comparable to firmware behaviour.
-
-- `--q-alt` — process noise on altitude (default `0.1`).
-- `--q-vel` — process noise on velocity (default `0.5`).
-- `--r-meas` — measurement noise variance (default `4.0`).
-- `--debounce` — apogee debounce ticks (default `3`).
-
-### Output
-
-- `--theme {light,dark,both}` — render plots matching the Furo Sphinx docs
-  theme (default `both`).
+- `--theme {light,dark,both}` — render plots matching the Furo Sphinx
+  docs theme (default `both`).
 - `--show` — open the plots in an interactive matplotlib window in
   addition to writing them to disk.
+- `--title TITLE` — override the plot title.
 
-On run, the script produces a graph and saves it to a file called
-`flight_simulation.png`. Here is an example run:
+Filter and attitude tuning is not exposed on the CLI on purpose: the
+constants at the top of the script (`CONFIG_FILTER_*`,
+`CONFIG_IMU_UP_AXIS_*`) are the same Kconfig defaults the firmware uses,
+so a default run is directly comparable to a default flight build. To
+sweep tuning, edit those constants or use
+[`sweep_apogee.py`](sweep_apogee.md).
 
-```bash
-Plot saved to flight_simulation.png
-Ground ref pressure: 101340 Pa (1013.4 hPa)
-True apogee:         500.0 m
-Filter apogee:       496.4 m (t = 12.56 s)
+The output file is `flight_simulation.png` (or `_light` / `_dark`
+suffixed when `--theme both` is used). A typical run looks like this:
+
+```
+Ground ref pressure: 101326 Pa (1013.3 hPa)
+True apogee:         500.0 m (t = 12.71 s)
+Filter apogee:       499.7 m (t = 12.78 s)
+Attitude g_mag:      9.794 m/s²
+Gated baro updates:  0
+[light] Plot saved to flight_simulation.png
 ```
 
 ```{image} /img/flight_simulation_light.png
@@ -90,15 +85,21 @@ Filter apogee:       496.4 m (t = 12.56 s)
 
 ## Reuse from other scripts
 
-`sim_flight_kalman.py` also exposes the helpers used by
-[`sweep_apogee.py`](sweep_apogee.md):
+The filter and attitude mirrors are exposed as plain module-level
+functions so they can be driven from any other Python tool that wants
+to replay or sweep them:
 
-- `parse_influx(path)` — parse a recorded `flights.influx` telemetry dump.
-- `parse_state_audit(path)` — parse a state-machine audit log.
-- `segment_flights(events)` — split a trace into `(boost_ms, end_ms)`
-  ranges on `ARMED → BOOST` transitions.
-- `process_real_flight(...)` — run one flight through the Python filter and
-  return a dict of time series plus detected apogee indices.
+- `Filter`, `filter_init`, `filter_predict`, `filter_update`,
+  `filter_detect_apogee`, `filter_votes`
+- `Attitude`, `attitude_init`, `attitude_calibrate_sample`,
+  `attitude_calibrate_finish`, `attitude_update`,
+  `attitude_is_calibrated`
+- `simulate_flight(...)` — generate the synthetic trajectory used by
+  the default run.
+
+Helpers for loading recorded telemetry (`parse_influx`,
+`parse_state_audit`, `segment_flights`, etc.) now live in
+[`plot_flight_data.py`](plot_flight_data.md).
 
 ## Requirements
